Move find_file to utils
